+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeteredData/V1/Model/ForwardMeteredDataAcceptedV1.cs
@@ -26,17 +26,8 @@ public record ForwardMeteredDataAcceptedV1(
     string MeteringPointId,
     MeteringPointType MeteringPointType,
     string ProductNumber,
-    MeasurementUnit MeasureUnit,
     DateTimeOffset RegistrationDateTime,
-    Resolution Resolution,
     DateTimeOffset StartDateTime,
     DateTimeOffset EndDateTime,
-    IReadOnlyCollection<ForwardMeteredDataAcceptedV1.AcceptedEnergyObservation> AcceptedEnergyObservations,
-    IReadOnlyCollection<MarketActorRecipientV1> MarketActorRecipients)
-    : IInputParameterDto
-{
-    public record AcceptedEnergyObservation(
-        int Position,
-        decimal? EnergyQuantity,
-        Quality? QuantityQuality);
-}
+    IReadOnlyCollection<ReceiversWithMeteredDataV1> ReceiversWithMeteredData)
+    : IInputParameterDto;